requirements.txt
streamlit
matplotlib
japanize_matplotlib
plotly.express
StandardScaler
scipy
KMeans